# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Project Management API
  version: 1.0.0
  description: A REST API for a Project Management application allowing teams to organize work, create tasks, and collaborate via comments.
servers:
  - url: http://localhost:3000
    description: Local development server
tags:
  - name: Auth
    description: Authentication endpoints
  - name: Projects
    description: Project management endpoints
  - name: Tasks
    description: Task management endpoints
  - name: Comments
    description: Task collaboration endpoints
  - name: Monitoring
    description: Operational endpoints

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Project:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Website Redesign"
        description:
          type: string
          example: "Revamping the corporate site"

    Task:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: "Design Homepage"
        status:
          type: string
          example: "Todo"
        projectId:
          type: integer
          example: 1
        assigneeId:
          type: integer
          example: 2

    Comment:
      type: object
      properties:
        id:
          type: integer
          example: 1
        content:
          type: string
          example: "I will start on this tomorrow."
        taskId:
          type: integer
          example: 1
        authorId:
          type: integer
          example: 2

paths:
  /health:
    get:
      summary: Health Check
      tags:
        - Monitoring
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                status: "ok"
                message: "Project Management API is running"

  /api/auth/signup:
    post:
      summary: Signup a new user
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email:
                  type: string
                  example: "user@example.com"
                password:
                  type: string
                  example: "securepassword123"
                name:
                  type: string
                  example: "John Doe"
      responses:
        "201":
          description: Created
          content:
            application/json:
              example:
                id: 1
                email: "user@example.com"
                name: "John Doe"
        "400":
          description: Bad request (Validation errors)
        "409":
          description: Email already in use

  /api/auth/login:
    post:
      summary: Authenticate user
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  example: "user@example.com"
                password:
                  type: string
                  example: "securepassword123"
      responses:
        "200":
          description: OK
          content:
            application/json:
              example:
                token: "eyJhbGciOiJIUzI1NiIsIn..."
        "400":
          description: Bad request
        "401":
          description: Invalid credentials

  /api/projects:
    post:
      summary: Create a project
      tags:
        - Projects
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  example: "Website Redesign"
                description:
                  type: string
                  example: "Revamping the corporate site"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "409":
          description: Duplicate project name

    get:
      summary: Get user's projects
      tags:
        - Projects
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
        "401":
          description: Unauthorized

  /api/projects/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      summary: Get project details
      tags:
        - Projects
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden (Not a member)
        "404":
          description: Not found

    put:
      summary: Update project
      tags:
        - Projects
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Website Redesign V2"
                description:
                  type: string
                  example: "Updated scope"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden (Insufficient permissions)
        "404":
          description: Not found

    delete:
      summary: Delete project
      tags:
        - Projects
      security:
        - bearerAuth: []
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found

  /api/tasks:
    post:
      summary: Create a task
      tags:
        - Tasks
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, projectId]
              properties:
                title:
                  type: string
                  example: "Design Homepage"
                status:
                  type: string
                  example: "Todo"
                projectId:
                  type: integer
                  example: 1
                assigneeId:
                  type: integer
                  example: 2
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
        "409":
          description: Duplicate task title

    get:
      summary: Get user's tasks
      tags:
        - Tasks
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        "401":
          description: Unauthorized

  /api/tasks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      summary: Get task details
      tags:
        - Tasks
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found

    put:
      summary: Update task
      tags:
        - Tasks
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                status:
                  type: string
                assigneeId:
                  type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found

    delete:
      summary: Delete task
      tags:
        - Tasks
      security:
        - bearerAuth: []
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found

  /api/comments:
    post:
      summary: Create comment
      tags:
        - Comments
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content, taskId]
              properties:
                content:
                  type: string
                  example: "I will start on this tomorrow."
                taskId:
                  type: integer
                  example: 1
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found

    get:
      summary: Get user's authored comments
      tags:
        - Comments
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
        "401":
          description: Unauthorized

  /api/comments/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      summary: Get comment details
      tags:
        - Comments
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found

    put:
      summary: Update comment
      tags:
        - Comments
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found

    delete:
      summary: Delete comment
      tags:
        - Comments
      security:
        - bearerAuth: []
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found